requirements.txt
google-genai
pydub
python-magic